+++ src/inet/common/ProtocolUtils.cc
@@ -45,7 +45,7 @@ const Protocol *popEncapsulationProtocolReq(Packet *packet)
 INET_API bool hasEncapsulationProtocolReq(Packet *packet, const Protocol *protocol)
 {
     if (auto encapsulationProtocolReq = packet->addTagIfAbsent<EncapsulationProtocolReq>()) {
-        for (int i = 0; i < encapsulationProtocolReq->getProtocolArraySize(); i++)
+        for (size_t i = 0; i < encapsulationProtocolReq->getProtocolArraySize(); i++)
             if (encapsulationProtocolReq->getProtocol(i) == protocol)
                 return true;
     }
@@ -55,7 +55,7 @@ INET_API bool hasEncapsulationProtocolReq(Packet *packet, const Protocol *protoc
 void removeEncapsulationProtocolReq(Packet *packet, const Protocol *protocol)
 {
     if (auto encapsulationProtocolReq = packet->addTagIfAbsent<EncapsulationProtocolReq>()) {
-        for (int i = 0; i < encapsulationProtocolReq->getProtocolArraySize(); i++) {
+        for (size_t i = 0; i < encapsulationProtocolReq->getProtocolArraySize(); i++) {
             if (encapsulationProtocolReq->getProtocol(i) == protocol) {
                 encapsulationProtocolReq->eraseProtocol(i);
                 break;
