+++ src/inet/common/TimeTag.h
@@ -17,9 +17,9 @@ template<typename T>
 void increaseTimeTag(const Ptr<Chunk>& chunk, simtime_t bitDuration, simtime_t packetDuration)
 {
     chunk->mapAllTagsForUpdate<T>(b(0), chunk->getChunkLength(), [&] (b offset, b length, T *timeTag) {
-        for (int i = 0; i < (int)timeTag->getBitTotalTimesArraySize(); i++)
+        for (size_t i = 0; i < timeTag->getBitTotalTimesArraySize(); i++)
             timeTag->setBitTotalTimes(i, timeTag->getBitTotalTimes(i) + bitDuration);
-        for (int i = 0; i < (int)timeTag->getPacketTotalTimesArraySize(); i++)
+        for (size_t i = 0; i < timeTag->getPacketTotalTimesArraySize(); i++)
             timeTag->setPacketTotalTimes(i, timeTag->getPacketTotalTimes(i) + packetDuration);
     });
 }
@@ -28,9 +28,9 @@ template<typename T>
 void increaseTimeTag(Packet *packet, simtime_t bitDuration, simtime_t packetDuration)
 {
     packet->mapAllRegionTagsForUpdate<T>(b(0), packet->getTotalLength(), [&] (b offset, b length, const Ptr<T>& timeTag) {
-        for (int i = 0; i < (int)timeTag->getBitTotalTimesArraySize(); i++)
+        for (size_t i = 0; i < timeTag->getBitTotalTimesArraySize(); i++)
             timeTag->setBitTotalTimes(i, timeTag->getBitTotalTimes(i) + bitDuration);
-        for (int i = 0; i < (int)timeTag->getPacketTotalTimesArraySize(); i++)
+        for (size_t i = 0; i < timeTag->getPacketTotalTimesArraySize(); i++)
             timeTag->setPacketTotalTimes(i, timeTag->getPacketTotalTimes(i) + packetDuration);
     });
 }
