+++ src/inet/linklayer/ieee80211/mac/aggregation/MpduAggregation.cc
@@ -19,7 +19,7 @@ Packet *MpduAggregation::aggregateFrames(std::vector<Packet *> *frames)
     EV_DEBUG << "Aggregating " << frames->size() << " packets into A-MPDU.\n";
     auto aggregatedFrame = new Packet();
     std::string aggregatedName;
-    for (int i = 0; i < (int)frames->size(); i++) {
+    for (size_t i = 0; i < frames->size(); i++) {
         auto mpduSubframeHeader = makeShared<Ieee80211MpduSubframeHeader>();
         auto frame = frames->at(i);
         auto mpdu = frame->peekAll();
